==> distance_methods-20221107T174152Z-001/distance_methods/source_code/notebooks/1_clustering_using_distances_global.ipynb <==
pip install python-louvain

import pandas as pd
import numpy as np
import community.community_louvain as community_louvain
import networkx as nx

dataset = {
    0 : "absorption",
    1 : "enantioselectivity",
    2 : "localization",
    3 : "T50"
}
method = {
    0:"FFT",
    1:"NLP",
    2:"Properties"
}
bioembedding = {
    0:"bepler",
    1:"esm",
    2:"fasttext",
    3:"plus_rnn",
    4:"prottrans"
}
distances = {
    1 : "Euclidean",
    2 : "Braycurtis",
    3 : "Canberra",
    4 : "Chebyshev",
    5 : "Cityblock",
    6 : "Correlation",
    7 : "Cosine",
    8 : "Minkowski",
    9 : "Hamming"
}

rf = [""]

for a in range(0,4):
    for b in range(0,3):
        if b == 0:
            for c in range(0,8):
                for d in range(1,10):
                    df_distances = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/distance_estimator.csv")
                    if d == 6:
                        for i in range(len(df_distances)):
                            df_distances.replace(to_replace=df_distances['distance'][i] ,value=abs(df_distances['distance'][i]))

                    q1 = np.quantile(df_distances['distance'], .25)
                    q3 = np.quantile(df_distances['distance'], .75)

                    IQR = q3 - q1
                    IQR = 1 * IQR

                    low_freq = q1 - IQR
                    high_freq = q3 + IQR

                    df_filter = df_distances.loc[df_distances['distance'] <= q3]
                    df_filter = df_filter.reset_index()

                    graph_data = nx.Graph()

                    id_list1 = df_filter['id_seq1'].to_list()
                    id_list2 = df_filter['id_seq2'].to_list()

                    nodos = id_list1 + id_list2
                    nodos = list(set(nodos))

                    for node in nodos:
                        graph_data.add_node(node)

                    for i in range(len(df_filter)):
                        graph_data.add_edge(df_filter['id_seq1'][i],
                                            df_filter['id_seq2'][i],
                                            weight=df_filter['distance'][i]
                        )
                    partition = community_louvain.best_partition(graph_data)
                    modularity_value = round(community_louvain.modularity(partition, graph_data),3)

                    res = "Modularidad de "+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+" es de "+str(modularity_value)
                    rf.append(res)
                    matrix_group = []

                    for element in partition:
                        row = [element, partition[element]]
                        matrix_group.append(row)
                    results = pd.DataFrame(matrix_group, columns=['id','label'])

                    results.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/graph_clustering_sequences.csv", index=False)

        if b == 1:
            for c in range(0,5):
                for d in range(1,10):
                    df_distances = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+"/distance_estimator.csv")
                    if d == 6:
                        for i in range(len(df_distances)):
                            df_distances.replace(to_replace=df_distances['distance'][i] ,value=abs(df_distances['distance'][i]))

                    q1 = np.quantile(df_distances['distance'], .25)
                    q3 = np.quantile(df_distances['distance'], .75)

                    IQR = q3 - q1
                    IQR = 1 * IQR

                    low_freq = q1 - IQR
                    high_freq = q3 + IQR

                    df_filter = df_distances.loc[df_distances['distance'] <= q3]
                    df_filter = df_filter.reset_index()

                    graph_data = nx.Graph()

                    id_list1 = df_filter['id_seq1'].to_list()
                    id_list2 = df_filter['id_seq2'].to_list()

                    nodos = id_list1 + id_list2
                    nodos = list(set(nodos))

                    for node in nodos:
                        graph_data.add_node(node)

                    for i in range(len(df_filter)):
                        graph_data.add_edge(df_filter['id_seq1'][i],
                                            df_filter['id_seq2'][i],
                                            weight=df_filter['distance'][i]
                        )
                    partition = community_louvain.best_partition(graph_data)
                    modularity_value = round(community_louvain.modularity(partition, graph_data),3)

                    res = "Modularidad de "+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+" es de "+str(modularity_value)
                    rf.append(res)

                    matrix_group = []

                    for element in partition:
                        row = [element, partition[element]]
                        matrix_group.append(row)
                    results = pd.DataFrame(matrix_group, columns=['id','label'])

                    results.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+"/graph_clustering_sequences.csv", index=False)

        if b == 2:
            for c in range(0,8):
                for d in range(1,10):
                    df_distances = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/distance_estimator.csv")
                    if d == 6:
                        for i in range(len(df_distances)):
                            df_distances.replace(to_replace=df_distances['distance'][i] ,value=abs(df_distances['distance'][i]))

                    q1 = np.quantile(df_distances['distance'], .25)
                    q3 = np.quantile(df_distances['distance'], .75)

                    IQR = q3 - q1
                    IQR = 1 * IQR

                    low_freq = q1 - IQR
                    high_freq = q3 + IQR

                    df_filter = df_distances.loc[df_distances['distance'] <= q3]
                    df_filter = df_filter.reset_index()

                    graph_data = nx.Graph()

                    id_list1 = df_filter['id_seq1'].to_list()
                    id_list2 = df_filter['id_seq2'].to_list()

                    nodos = id_list1 + id_list2
                    nodos = list(set(nodos))

                    for node in nodos:
                        graph_data.add_node(node)

                    for i in range(len(df_filter)):
                        graph_data.add_edge(df_filter['id_seq1'][i],
                                            df_filter['id_seq2'][i],
                                            weight=df_filter['distance'][i]
                        )
                    partition = community_louvain.best_partition(graph_data)
                    modularity_value = round(community_louvain.modularity(partition, graph_data),3)

                    res = "Modularidad de "+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+" es de "+str(modularity_value)
                    rf.append(res)
                    matrix_group = []

                    for element in partition:
                        row = [element, partition[element]]
                        matrix_group.append(row)
                    results = pd.DataFrame(matrix_group, columns=['id','label'])

                    results.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/graph_clustering_sequences.csv", index=False)

rf

df_filter

partition = community_louvain.best_partition(graph_data)
modularity_value = round(community_louvain.modularity(partition, graph_data), 3)

print("Modularity: ", modularity_value)

#obtenemos los grupos
