==> distance_methods-20221107T174152Z-001/distance_methods/source_code/notebooks/3_merge_global.ipynb <==
import pandas as pd

dataset = {
    0 : "absorption",
    1 : "enantioselectivity",
    2 : "localization",
    3 : "T50"
}
method = {
    0:"FFT",
    1:"NLP",
    2:"Properties"
}
bioembedding = {
    0:"bepler",
    1:"esm",
    2:"fasttext",
    3:"plus_rnn",
    4:"prottrans"
}
distances = {
    1 : "Euclidean",
    2 : "Braycurtis",
    3 : "Canberra",
    4 : "Chebyshev",
    5 : "Cityblock",
    6 : "Correlation",
    7 : "Cosine",
    8 : "Minkowski",
    9 : "Hamming"
}

for a in range(0,4):
    for b in range(0,3):
        if b == 0:
            for c in range(0,8):
                clust_kmeans = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_kmeans.csv")
                clust_birch = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_birch.csv")
                clust_aglo = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_aglomerative.csv")
                clust_meanshift = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_meanshift.csv")
                clust_affinity = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_affinity.csv")

                clust_merge = clust_kmeans

                profiles = []

                for index in clust_merge.index:
                    profile = "{}-{}-{}-{}-{}".format(
                        clust_kmeans['labels'][index],
                        clust_birch['labels'][index],
                        clust_aglo['labels'][index],
                        clust_meanshift['labels'][index],
                        clust_affinity['labels'][index]
                    )
                    profiles.append(profile)

                clust_merge['labels'] = profiles
                clust_merge['labels'].unique()

                clust_merge = clust_merge.drop(columns=['Unnamed: 0'])
                clust_merge.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences.csv")

                for d in range(1,10):

                    cluster_graph = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/graph_clustering_sequences.csv")
                    cluster_unsupervised = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences.csv")

                    df_merge = cluster_graph.merge(cluster_unsupervised, left_on='id',right_on='id')
                    df_merge = df_merge.drop(columns=['Unnamed: 0'])

                    profiles = []

                    for index in df_merge.index:
                        profile = "{}-{}".format(
                            df_merge['label'][index],
                            df_merge['labels'][index],
                        )
                        profiles.append(profile)

                    df_merge['profile'] = profiles
                    df_merge['profile'].unique()

                    df_merge.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/generated_profiles.csv",index=False)

        if b == 1:
            for c in range(0,5):
                clust_kmeans = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences_kmeans.csv")
                clust_birch = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences_birch.csv")
                clust_aglo = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences_aglomerative.csv")
                clust_meanshift = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences_meanshift.csv")
                clust_affinity = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences_affinity.csv")
                clust_merge = clust_kmeans

                profiles = []

                for index in clust_merge.index:
                    profile = "{}-{}-{}-{}-{}".format(
                        clust_kmeans['labels'][index],
                        clust_birch['labels'][index],
                        clust_aglo['labels'][index],
                        clust_meanshift['labels'][index],
                        clust_affinity['labels'][index]
                    )
                    profiles.append(profile)

                clust_merge['labels'] = profiles
                clust_merge['labels'].unique()

                clust_merge = clust_merge.drop(columns=['Unnamed: 0'])
                clust_merge.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences.csv")

                for d in range(1,10):

                    cluster_graph = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+"/graph_clustering_sequences.csv")
                    cluster_unsupervised = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/unsupervised_clustering_sequences.csv")

                    df_merge = cluster_graph.merge(cluster_unsupervised, left_on='id',right_on='id')
                    df_merge = df_merge.drop(columns=['Unnamed: 0'])

                    profiles = []

                    for index in df_merge.index:
                        profile = "{}-{}".format(
                            df_merge['label'][index],
                            df_merge['labels'][index],
                        )
                        profiles.append(profile)

                    df_merge['profile'] = profiles
                    df_merge['profile'].unique()

                    df_merge.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/"+bioembedding[c]+"/"+distances[d]+"/generated_profiles.csv",index=False)

        if b == 2:
            for c in range(0,8):
                clust_kmeans = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_kmeans.csv")
                clust_birch = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_birch.csv")
                clust_aglo = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_aglomerative.csv")
                clust_meanshift = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_meanshift.csv")
                clust_affinity = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences_affinity.csv")
                clust_merge = clust_kmeans

                profiles = []

                for index in clust_merge.index:
                    profile = "{}-{}-{}-{}-{}".format(
                        clust_kmeans['labels'][index],
                        clust_birch['labels'][index],
                        clust_aglo['labels'][index],
                        clust_meanshift['labels'][index],
                        clust_affinity['labels'][index]
                    )
                    profiles.append(profile)

                clust_merge['labels'] = profiles
                clust_merge['labels'].unique()

                clust_merge = clust_merge.drop(columns=['Unnamed: 0'])
                clust_merge.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences.csv")

                for d in range(1,10):

                    cluster_graph = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/graph_clustering_sequences.csv")
                    cluster_unsupervised = pd.read_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/unsupervised_clustering_sequences.csv")

                    df_merge = cluster_graph.merge(cluster_unsupervised, left_on='id',right_on='id')
                    df_merge = df_merge.drop(columns=['Unnamed: 0'])

                    profiles = []

                    for index in df_merge.index:
                        profile = "{}-{}".format(
                            df_merge['label'][index],
                            df_merge['labels'][index],
                        )
                        profiles.append(profile)

                    df_merge['profile'] = profiles
                    df_merge['profile'].unique()

                    df_merge.to_csv("../../results_demo/"+dataset[a]+"/"+method[b]+"/Group_"+str(c)+"/"+distances[d]+"/generated_profiles.csv",index=False)
